+++ packages/tswebextension/src/lib/mv3/background/index.ts
@@ -17,21 +17,19 @@ export {
     UnavailableFilterSourceError,
     UnavailableRuleSetSourceError,
     ResourceType,
+    RULESET_NAME_PREFIX,
 } from '@adguard/tsurlfilter/es/declarative-converter';
 
+// re-exports to prevent collision, when both tsurlfilter and tswebextension are imported
 export {
     getRuleSourceIndex,
     getRuleSourceText,
     type PreprocessedFilterList,
     FilterListPreprocessor,
     preprocessedFilterListValidator,
-} from '@adguard/tsurlfilter';
-
-// re-exports to prevent collision, when both tsurlfilter and tswebextension are imported
-export type {
-    NetworkRule,
-    CosmeticRule,
-    NetworkRuleOption,
+    type NetworkRule,
+    type CosmeticRule,
+    type NetworkRuleOption,
 } from '@adguard/tsurlfilter';
 
 export * from './app';
@@ -45,7 +43,6 @@ export type {
     SettingsConfigMV3 as SettingsConfig,
 } from './configuration';
 export { MessageType } from '../../common/message-constants';
-export { RULESET_NAME_PREFIX } from '@adguard/tsurlfilter/es/declarative-converter';
 export { companiesDbService } from '../../common/companies-db-service';
 export {
     getDomain,